+++ src/components/OutputBox.jsx
