+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
 import {Provider} from "react-redux";
-import store from './redux/store';
+import store from './store/store';
 
 ReactDOM.render(
   <Provider store={store}>
